+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Released: [Place date of release here]
 
 ### New Features
 
+* `faspex5`: Package browsing now available for all inbox types.
+
 ### Issues Fixed
 
 * `aoc`: Public short links can now be modified.
@@ -17,7 +19,6 @@ Released: [Place date of release here]
 * `config`: Deprecated and removed action `ascp connect`.
 * **global**: Deprecated and removed option `property` made redundant with `@:` and dotted values. Instead of `--property=foo bar`, use `@: foo=bar`.
 * `faspex5`: Default value for option `box` is now `inbox_all` to match web UI, previously: `inbox`.
-* `faspex5`: Package browsing now available for all inbox types.
 
 ## 4.25.2